check for file existing before remove it

+++ upload/system/library/cache.php
@@ -41,13 +41,13 @@ final class Cache {
         $watermark   = DIR_IMAGE . 'common' . DIR_SEPARATOR . 'watermark.png';
         $cached_url  = URL_BASE . 'image' . DIR_SEPARATOR . 'cache' . DIR_SEPARATOR . $user_id . DIR_SEPARATOR . $name . '-' . (int) $best_fit . '-' . $width . '-' . $height . '.' . STORAGE_IMAGE_EXTENSION;
 
+        // Force reset
+        if ($overwrite && file_exists($overwrite)) {
+            unlink($cache);
+        }
+
         // If image is cached
         if (file_exists($cache)) {
-
-            // Force reset
-            if ($overwrite) {
-                unlink($cache);
-            }
 
             return $cached_url;
 
@@ -90,18 +90,20 @@ final class Cache {
     * @return string Cached Audio URL
     */
     public function audio($name, $user_id, $bit_rate = 320, $overwrite = true, $start = false, $end = false) {
-
+        ini_set('display_errors', 1);
+        ini_set('display_startup_errors', 1);
+        error_reporting(E_ALL);
         $storage     = DIR_STORAGE . $user_id . DIR_SEPARATOR . $name . '.' . STORAGE_AUDIO_EXTENSION;
         $cache       = DIR_AUDIO . 'cache' . DIR_SEPARATOR . $user_id . DIR_SEPARATOR . $name . '.mp3'; // todo
         $cached_url  = URL_BASE . 'audio' . DIR_SEPARATOR . 'cache' . DIR_SEPARATOR . $user_id . DIR_SEPARATOR . $name . '.mp3'; // todo
 
+        // Force reset
+        if ($overwrite && file_exists($overwrite)) {
+            unlink($cache);
+        }
+
         // If audio is cached
         if (file_exists($cache)) {
-
-            // Force reset
-            if ($overwrite) {
-                unlink($cache);
-            }
 
             return $cached_url;
